fix(reindex): Don't leave INVALID indexes around
Sometimes index creation can fail, leaving the index in an INVALID
state. Previously this software would continue happily along. This
essentially mean the index was gone, since INVALID indexes can't be
used.

Instead it will try to recreate the index maximum 3 times, and if that
doesn't work, it'll put the old, bloated but working, index back.

This should prevent it leaving INVALID indexes.